neuralmt/requirements.txt
numpy
notebook
jupyter_contrib_nbextensions
jupyter_nbextensions_configurator
torch
matplotlib>=2.0.2
tqdm>=4.19.5
torchtext>=0.2.1
Pillow>=5.0.0
seaborn>=0.8.1
tensorboardX>=1.1
pandas
sacrebleu
